requirements.txt
requests>=2.32.0
questionary>=2.1.0
Pillow>=10.0.0
rich>=13.0.0